+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Element injection into options nested in repeatable sets
+- 
 ### Changed
 
 - `pointCadastralParcel` to only show results grid if address has `features`